old-tests/submission/Microsoft/canvas/canvas_pixelManipulation_imagedata_002.htm

<!doctype HTML> 
<html>
    <head>
        <title>HTML5 Canvas Test:  CanvasPixelArray is a readonly attribute of ImageData</title>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com" />
        <link rel="help" href="http://www.w3.org/TR/2dcontext/#imagedata" />
        <meta name="assert" content="CanvasPixelArray is a readonly attribute of ImageData." />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
      <script type="text/javascript">
           "use strict";
           setup({explicit_done: true});
           function runTest() 
            {
                test(function ()
                {
                var canvas = document.getElementById("canvas1");
                var ctx = canvas.getContext("2d");

                // Get some image data.
                var imgData = ctx.getImageData(0, 0, 1, 1);
                var pixelArray = imgData.data;
                imgData.data[0] = 10;
                pixelArray[0] = 20;
                assert_readonly(imgData, "data");
                assert_throws(new TypeError(), function() { imgData.data = pixelArray; });
                }, "CanvasPixelArray is a readonly attribute of ImageData.");
                done();
            }
        </script>
    </head>
    <body onload="runTest()">
        <p>Description: CanvasPixelArray is a readonly attribute of ImageData.</p>
	<div id='log'></div>
        <canvas id="canvas1" width="300" height="150">Browser does not support HTML5 Canvas.</canvas>
    </body>
</html>
